requirements.txt
flask
langchain
langchain-openai
langchain-community
langchain-chroma
python-dotenv